+++ addon/frontend/src/main.tsx
@@ -1,16 +1,16 @@
+import { MantineProvider } from "@mantine/core";
 import "@mantine/core/styles.css";
 import "@mantine/dates/styles.css";
-import "@mantine/notifications/styles.css";
-import { MantineProvider } from "@mantine/core";
 import { Notifications } from "@mantine/notifications";
+import "@mantine/notifications/styles.css";
 import ReactDOM from "react-dom/client";
 import App from "./App";
 import AppErrorBoundary from "./components/AppErrorBoundary";
 import "./index.css";
 
 const theme = {
-  primaryColor: "green",
-  fontFamily: "Roboto, Arial, sans-serif",
+  primaryColor: "blue",
+  fontFamily: "Noto Sans, Arial, sans-serif",
 };
 
 ReactDOM.createRoot(document.getElementById("root")!).render(